+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - added :tf-map-handler-fn to support pruning of the terms for tfidf
   - metamorph'ed  bow->tfidf reuses tf-map from :fit in :transform
   - 2 weighting schees for tf calculation
+  - 3 weighting schemes for idf calculation
 
 
 ## v6.2.585